+++ routes/_route.js
@@ -10,7 +10,11 @@ module.exports = function( app ) {
   app.use( '/process', require( './process' ) );
   app.use( '/usage', require( './usage' ) );
   app.use( '/error', require( './error.js' ) );
-  app.use( '/log', require( './log.js' ) );
+
+  if ( process.env.LOG_ROUTE ) {
+    app.use( '/' + process.env.LOG_ROUTE, require( './log.js' ) );
+  }
+
   // var usersRouter = require('./users'); // @deprecated
 
   // Favicon